requirements.txt
apscheduler
scikit-learn
pandas
alabaster==0.7.11
Babel==2.6.0
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-bootstrap4==0.0.6
django-crispy-forms==1.7.2
django-registration-redux==2.4
django-search-listview==0.2.1
django-sslserver==0.20
docutils==0.14
google-api-core==1.2.1
googleapis-common-protos==1.5.3
grpcio==1.13.0
idna==2.7
imagesize==1.0.0
Jinja2==2.10
MarkupSafe==1.0
packaging==17.1
protobuf==3.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
Pygments==2.2.0
pyparsing==2.2.0
pytz==2018.4
requests==2.19.1
rsa==3.4.2
selenium==3.13.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinx-rtd-theme==0.4.0
sphinxcontrib-websupport==1.1.0
urllib3==1.23
